+++ components/collapsible/collapsible.vue
@@ -12,9 +12,6 @@
         'd-dt-collapsibe__anchor',
         anchorClass,
       ]"
-      @click.capture="defaultToggleOpen"
-      @keydown.enter="defaultToggleOpen"
-      @keydown.space="defaultToggleOpen"
     >
       <!-- @slot Slot for the anchor element that toggles the collapsible content -->
       <slot
@@ -30,9 +27,11 @@
           kind="muted"
           :aria-controls="id"
           :aria-expanded="`${isOpen}`"
+          :aria-label="ariaLabel"
           :style="{
             'width': maxWidth,
           }"
+          @click="defaultToggleOpen"
         >
           <dt-icon
             :name=" isOpen ? 'chevron-down' : 'chevron-right'"
@@ -50,10 +49,9 @@
     <dt-collapsible-lazy-show
       :id="id"
       ref="contentWrapper"
-      :aria-hidden="`${!isOpen}`"
+      :aria-hidden="`${!isOpen && !hasContentOnCollapse}`"
       :aria-labelledby="labelledBy"
-      :aria-label="ariaLabel"
-      :show="isOpen"
+      :is-expanded="isOpen"
       :element-type="contentElementType"
       :class="[
         'd-dt-collapsible__content',
@@ -66,13 +64,19 @@
       tabindex="-1"
       appear
       v-on="$listeners"
-      @after-leave="onLeaveTransitionComplete"
-      @after-enter="onEnterTransitionComplete"
     >
-      <!-- @slot Slot for the collapsible element that is expanded by the anchor -->
-      <slot
-        name="content"
-      />
+      <template #contentOnExpanded>
+        <!-- @slot Slot for content that is shown when collapsible is expanded -->
+        <slot
+          name="contentOnExpanded"
+        />
+      </template>
+      <template #contentOnCollapsed>
+        <!-- @slot Slot for content that is shown when collapsible is collapsed -->
+        <slot
+          name="contentOnCollapsed"
+        />
+      </template>
     </dt-collapsible-lazy-show>
   </component>
 </template>
@@ -179,7 +183,7 @@ export default {
     },
 
     /**
-     * Label on the collapsible content. Should provide this or ariaLabelledBy but not both.
+     * Label on the anchor. Should provide this or ariaLabelledBy but not both.
      */
     ariaLabel: {
       type: String,
@@ -222,14 +226,11 @@ export default {
     labelledBy () {
       // aria-labelledby should be set only if aria-labelledby is passed as a prop, or if
       // there is no aria-label and the labelledby should point to the anchor
-      return this.ariaLabelledby || (!this.ariaLabel && getUniqueString('DtCollapsible__anchor'));
+      return this.ariaLabelledBy || (!this.ariaLabel && getUniqueString('DtCollapsible__anchor'));
     },
 
-    contentStyle () {
-      return {
-        'max-height': this.maxHeight,
-        'max-width': this.maxWidth,
-      };
+    hasContentOnCollapse () {
+      return !!this.$slots.contentOnCollapsed;
     },
   },
 
@@ -245,29 +246,25 @@ export default {
     },
   },
 
-  methods: {
-    onLeaveTransitionComplete () {
-      this.$emit('opened', false);
-      if (this.open !== null) {
-        this.$emit('update:open', false);
-      }
-    },
+  created () {
+    this.validateProperAnchor();
+  },
 
-    onEnterTransitionComplete () {
-      this.$emit('opened', true, this.$refs.content);
-      if (this.open !== null) {
-        this.$emit('update:open', true);
-      }
-    },
+  methods: {
 
     defaultToggleOpen () {
+      this.$emit('opened', !this.isOpen);
       if (this.open === null) {
-        this.toggleOpen();
+        this.isOpen = !this.isOpen;
+      } else {
+        this.$emit('update:open', !this.isOpen);
       }
     },
 
-    toggleOpen () {
-      this.isOpen = !this.isOpen;
+    validateProperAnchor () {
+      if (!this.anchorText && !this.$scopedSlots.anchor) {
+        console.error('anchor text and anchor slot content cannot both be falsy');
+      }
     },
   },
 };